refactor: consolidate widget workbench modules
