+++ apps/nextjs/src/app/[locale]/manage/custom-widgets/packages/_package-collectors.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+import { extractErrorMessage } from "@homarr/common";
+
 import { useState } from "react";
 import {
   Alert,
@@ -13,12 +15,12 @@ import {
   Text,
   Textarea,
   TextInput,
+  Accordion,
+  Table,
 } from "@mantine/core";
-
 import type { RouterOutputs } from "@homarr/api";
 import { clientApi } from "@homarr/api/client";
 import { useI18n } from "@homarr/translation/client";
-import { PackageHistorySummary } from "./_package-history-summary";
 
 export function PackageCollectors({ itemId, queries }: { itemId: string; queries: string[] }) {
   const t = useI18n("customWidget.package");
@@ -61,7 +63,7 @@ export function PackageCollectors({ itemId, queries }: { itemId: string; queries
         },
       );
     } catch (cause) {
-      setError(cause instanceof Error ? cause.message : String(cause));
+      setError(extractErrorMessage(cause));
     }
   };
   return (
@@ -194,4 +196,72 @@ function CollectorRow({
       </Stack>
     </Paper>
   );
+}
+
+type History = RouterOutputs["widget"]["customApi"]["packageHistory"];
+
+function PackageHistorySummary({ data }: { data: History }) {
+  const t = useI18n("customWidget.package.historySummary");
+  const latest = data.samples.at(-1);
+  return (
+    <Stack gap="xs">
+      {latest && (
+        <Group justify="space-between">
+          <Text fw={600}>
+            {formatCollectedValue(latest.value)} {data.unit}
+          </Text>
+          <Text size="xs" c="dimmed">
+            {t("latest", { time: new Date(latest.timestamp).toLocaleString() })}
+          </Text>
+        </Group>
+      )}
+      {!latest && (
+        <Text size="sm" c="dimmed">
+          {t("empty")}
+        </Text>
+      )}
+      {"lastError" in data && data.lastError && <Alert color="yellow">{data.lastError}</Alert>}
+      <Accordion multiple>
+        <Accordion.Item value="samples">
+          <Accordion.Control>{t("samples", { count: data.samples.length })}</Accordion.Control>
+          <Accordion.Panel>
+            <div style={{ maxHeight: 220, overflow: "auto" }}>
+              <Table>
+                <Table.Thead>
+                  <Table.Tr>
+                    <Table.Th>{t("time")}</Table.Th>
+                    <Table.Th>{t("value")}</Table.Th>
+                  </Table.Tr>
+                </Table.Thead>
+                <Table.Tbody>
+                  {data.samples.toReversed().map((sample, index) => (
+                    <Table.Tr key={`${sample.timestamp}-${index}`}>
+                      <Table.Td>{new Date(sample.timestamp).toLocaleString()}</Table.Td>
+                      <Table.Td>
+                        {formatCollectedValue(sample.value)} {data.unit}
+                      </Table.Td>
+                    </Table.Tr>
+                  ))}
+                </Table.Tbody>
+              </Table>
+            </div>
+          </Accordion.Panel>
+        </Accordion.Item>
+        <Accordion.Item value="technical">
+          <Accordion.Control>{t("details")}</Accordion.Control>
+          <Accordion.Panel>
+            <Text component="pre" size="xs" style={{ whiteSpace: "pre-wrap", maxHeight: 200, overflow: "auto" }}>
+              {JSON.stringify(data, null, 2)}
+            </Text>
+          </Accordion.Panel>
+        </Accordion.Item>
+      </Accordion>
+    </Stack>
+  );
+}
+
+function formatCollectedValue(value: unknown): string {
+  if (typeof value === "number") return value.toLocaleString();
+  if (typeof value === "string") return value.slice(0, 240);
+  return (JSON.stringify(value) ?? "—").slice(0, 240);
 }
