+++ apps/nextjs/src/app/[locale]/manage/custom-widgets/packages/preview/[previewId]/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 import { Container } from "@mantine/core";
 import { auth } from "@homarr/auth/next";
-import { PackageSessionPreview } from "../../_package-session-preview";
+import { PackageSessionPreview } from "../../_package-preview";
 
 export default async function PackagePreviewPage({ params }: { params: Promise<{ previewId: string }> }) {
   const session = await auth();
