+++ packages/common/src/browser.ts
